# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "getAuthorList",
                "description": "Get all Authors",
                "operationId": "dfa7fd6f91da42a5d4b0173c5afd0d6b",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Author"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "createAuthor",
                "description": "Create Author",
                "operationId": "5637b607c8bfe7a1ff0b4ad336510890",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "getAuthorItem",
                "description": "Get Author",
                "operationId": "71001ad4c874760c093e03e936858bc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Author"
                ],
                "summary": "updateAuthor",
                "description": "Update Author",
                "operationId": "195f792922398f74a961edaa5e67e44d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Author"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Author"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Author"
                ],
                "summary": "deleteAuthor",
                "description": "Delete Author",
                "operationId": "65f9011398591ec301420c271e96e05b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "created_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}